fix(citizen): harmonyscore back to int

+++ drone_program/src/main/scala/DroneGenerator.scala
@@ -32,7 +32,7 @@ object DroneGenerator {
     case class Drone(id: Int, location: List[Double]) {
         
         // Method to schedule script execution every minute
-        def scheduleScriptExecution(executorService: ScheduledExecutorService, citizenList: List[(String, Double, Double, Double)], originTimestamp: Instant)(implicit ec: ExecutionContext): Future[Unit] = {
+        def scheduleScriptExecution(executorService: ScheduledExecutorService, citizenList: List[(String, Double, Double, Int)], originTimestamp: Instant)(implicit ec: ExecutionContext): Future[Unit] = {
             val promise = Promise[Unit]()
             // Schedule the script execution every minute
             executorService.scheduleAtFixedRate(
